feat(suite-native): only non-interactive items trigger
