+++ suite-native/module-accounts-management/src/components/AccountDetailCryptoValue.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 
-import { HStack } from '@suite-native/atoms';
+import { DiscreetTextTrigger, HStack } from '@suite-native/atoms';
 import { CryptoAmountFormatter } from '@suite-native/formatters';
 import { CryptoIcon } from '@suite-common/icons';
 import { NetworkSymbol } from '@suite-common/wallet-config';
@@ -15,12 +15,14 @@ export const AccountDetailCryptoValue = memo(
     ({ value, networkSymbol, isBalance = true }: AccountDetailBalanceProps) => (
         <HStack spacing="small" flexDirection="row" alignItems="center" justifyContent="center">
             <CryptoIcon symbol={networkSymbol} size="extraSmall" />
-            <CryptoAmountFormatter
-                value={value}
-                network={networkSymbol}
-                isBalance={isBalance}
-                adjustsFontSizeToFit
-            />
+            <DiscreetTextTrigger>
+                <CryptoAmountFormatter
+                    value={value}
+                    network={networkSymbol}
+                    isBalance={isBalance}
+                    adjustsFontSizeToFit
+                />
+            </DiscreetTextTrigger>
         </HStack>
     ),
 );
